+++ imports/ui/App.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { Hello } from './Hello.jsx';
-import { Info } from './Info.jsx';
-import { NavBar } from './NavBar.jsx';
-import { ExampleChart } from './ExampleChart.jsx';
+import { Hello } from './components/Hello.jsx';
+import { Info } from './components/Info.jsx';
+import { NavBar } from './components/NavBar.jsx';
+import { ExampleChart } from './components/ExampleChart.jsx';
 
 export const App = () => (
-  <div>
-    <NavBar/>
-    <Hello/>
-    <Info/>
-    <ExampleChart/>
-    
-  </div>
+    <div>
+        <NavBar />
+        <Hello />
+        <Info />
+        <ExampleChart />
+
+    </div>
 );
